Add session filter/backend, and filter/service prototypes
This adds session primitives and tests. This also has mock in-memory
store implementations of data stores for Secrets, Sessions, and Tokens.
It also includes JSON encoding/decoding for Tokens.

The prototypes of filters/sessions come from wkimeria's branch,
including them in the initial merge of this code into the main repo so
we can work off of the same code base.