Add plugabble auth and example app
This adds the foundation for our pluggable authentication API in the
borderpatrol.auth package. Further work is needed to flesh out an OAuth2
implementation in the example application, but there is a BasicAuth

This utilizes the new pluggable auth module to show how a basic auth
service/filter would be created.