Improve auth/session components
The old auth and session components were not very generic, which would
make this a less effective library for users. This is an attempt to
allow different types of sessions and authentication possible via
TypeClasses. Default implementation for JSON and Buf are included

Further, the interface for generating sessions is now a Future, because
entropy is a blocking operation.