Update README with up-to-date information
The README was in much need of info, this adds a much more detailed
explanation of things and instructions on how to run the example server.

This also includes an update to the build.sbt that disables Jmh since it
causes the example app to not run.

Part of the work required for #45