Add CSRF protection filter
This commit provides upstream services with a way to do CSRF without
needing to be stateful themselves. Instead, they need only communicate
with Border Patrol via two options:
 - Header (useful for javascript single page apps)
 - Query parameter (useful for all others)

Upon successful login, the user should be presented with a CSRF cookie
whos value upstreams will use to inject into requests for Border Patrol
to verify that there is no CSRF vulnerability. This pattern is known as
Double Submit Cookie.

The request will be routed through Border Patrol which will inspect both
the header or param, and the cookie value that each matches. Since we
use the capabilities of the SessionId, we also get the cryptographic
attributes in this CSRF cookie.

Once Border Patrol has verified the incoming CSRF header/param and
cookie it will inject a "verified" header with a configurable value (the
CsrfFilter just calls .toString on a Boolean)

The values for these headers are configurable, but have sane defaults
that shouldn't cause conflicts in services or existing headers or
parameters.

Fixes #48